+++ docs/.vitepress/utils/index.ts
@@ -1,0 +1,11 @@
+import { utoa } from './utils'
+const MAIN_FILE_NAME = 'App.vue'
+
+export const getPlaygroundEncoded = (source: string) => {
+  const code = decodeURIComponent(source)
+  const originCode = {
+    [MAIN_FILE_NAME]: code
+  }
+  const encoded = utoa(JSON.stringify(originCode))
+  return encoded
+}