+++ packages/components/table/src/index.vue
@@ -3,11 +3,11 @@
     <PlusTableTitleBar
       v-if="titleBar"
       :columns="columns"
-      :sub-columns="subColumns"
       :default-size="size"
       :title-bar="titleBar"
       @click-density="handleClickDensity"
       @filter-table="handleFilterTableConfirm"
+      @refresh="handleRefresh"
     >
       <template #title>
         <slot name="title" />
@@ -176,6 +176,7 @@ export interface PlusTableEmits {
   (e: 'clickActionConfirmCancel', data: ButtonsCallBackParams): void
   (e: 'dragSortEnd', newIndex: number, oldIndex: number): void
   (e: 'formChange', data: { value: any; prop: string; row: any; index: number; column: any }): void
+  (e: 'refresh'): void
 }
 
 defineOptions({
@@ -297,6 +298,11 @@ const handleClickDensity = (data: ComponentSize) => {
 
 const handleDragSortEnd = (newIndex: number, oldIndex: number) => {
   emit('dragSortEnd', newIndex, oldIndex)
+}
+
+// 刷新
+const handleRefresh = () => {
+  emit('refresh')
 }
 
 const handleFormChange = (data: {